+++ screens/product/section-two/index.js
@@ -68,7 +68,7 @@ export default function SectionTwo({ data, isDesktop }) {
                 <div
                     className={`hero-text ${
                         data.extraImage ? 'with-image' : ''
-                    } `}
+                    } ${data?.isLeftText && isDesktop ? 'left' : ''}`}
                 >
                     <h2 style={{ color: data.color }}>{data?.title ?? ''}</h2>
                     <div>{data?.description ?? ''}</div>